feat(users, repository service, REST API): Add user CRUD operations, refactor repository service, and implement password hashing
- Added storage, update, delete, and get operations for users in the database.
- Refactored the repository service to split into a base service, with each entity having its own part.
- Added REST API endpoints to store, update, delete, and get users from the database.
- Implemented password hashing and storage for each user.
